+++ lib/src/common/network/api_client.dart
@@ -1,30 +1,17 @@
 import 'dart:convert';
-import 'dart:typed_data';
 
 import 'package:flutter/foundation.dart';
 import 'package:http/http.dart';
-import 'package:http_parser/http_parser.dart';
-import 'package:wiredash/src/common/network/data_state.dart';
-
-typedef DataStateCallback<T> = void Function(DataState<T> dataState);
 
 class ApiClient {
-  ApiClient(
-      {@required this.httpClient,
-      @required this.projectId,
-      @required this.secret});
+  ApiClient({
+    @required this.httpClient,
+    @required this.projectId,
+    @required this.secret,
+  });
 
-  static const String _host = 'https://api.wiredash.io/';
-  static const String _feedbackPath = 'feedback';
-
-  static const String _parameterDeviceInfo = 'deviceInfo';
-  static const String _parameterEmail = 'email';
-  static const String _parameterPayload = 'payload';
-  static const String _parameterUser = 'user';
-
-  static const String _parameterFeedbackMessage = 'message';
-  static const String _parameterFeedbackScreenshot = 'file';
-  static const String _parameterFeedbackType = 'type';
+  static const String _host =
+      'http://localhost:3000/'; //'https://api.wiredash.io/';
 
   final Client httpClient;
   final String projectId;
@@ -88,55 +75,6 @@ class ApiClient {
       return json.decode(responseString) as Map<String, dynamic>;
     } catch (exception) {
       throw Exception('${exception.toString()}\n$responseString');
-    }
-  }
-
-  Future<void> sendFeedback({
-    @required Map<String, dynamic> deviceInfo,
-    String email,
-    @required String message,
-    Map<String, dynamic> payload,
-    Uint8List picture,
-    @required String type,
-    String user,
-    @required DataStateCallback<Map<String, dynamic>> onDataStateChanged,
-  }) async {
-    onDataStateChanged(
-      DataState<Map<String, dynamic>>.loading(),
-    );
-
-    MultipartFile screenshotFile;
-
-    if (picture != null) {
-      screenshotFile = MultipartFile.fromBytes(
-        _parameterFeedbackScreenshot,
-        picture,
-        filename: 'file',
-        contentType: MediaType('image', 'png'),
-      );
-    }
-
-    try {
-      onDataStateChanged(
-        DataState<Map<String, dynamic>>.success(
-          await post(
-            urlPath: '$_feedbackPath',
-            arguments: {
-              _parameterDeviceInfo: json.encode(deviceInfo),
-              if (email != null) _parameterEmail: email,
-              _parameterFeedbackMessage: message,
-              if (payload != null) _parameterPayload: json.encode(payload),
-              _parameterFeedbackType: type,
-              if (user != null) _parameterUser: user
-            },
-            files: [screenshotFile],
-          ),
-        ),
-      );
-    } catch (exception) {
-      onDataStateChanged(
-        DataState<Map<String, dynamic>>.error(exception),
-      );
     }
   }
 }