+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [0.0.2] - Confidential 👀 & Provider 🏗
+Wiredash now uses the Provider package for internal state management and supports version 3.0.0 and higher. If you are
+also using Provider in your app, please make sure to at least use version 3.0.0.
+
+* Added a Confidential widget to automatically hide sensitive widgets during screen capture
+* Added a ValueListener to check if Wiredash is in screen capture mode
+* Improved error handling when there is no valid root navigator key
+* Minor bug fixes
+
 ## [0.0.1] - Public Release
 
 * Wiredash gets released to the public 🎉